move utils package to root
